requirements.txt
celery==5.1.2
content==0.0.1
decorator==5.1.0
Django==3.2.6
django-celery-results==2.2.0
django-ckeditor==6.1.0
django-debug-toolbar==3.2.2
password==0.2
Pillow==8.3.1
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
python-dotenv==0.19.1
view==0.1
